moved resource classes to SPI
